refactoring tests out of test_initial.py
